Change image data arrays to use 16bit, don't feel like doing manual conversions. Will just do in-code

+++ Starship_GalaxyWars/include/Entity.h
@@ -9,7 +9,7 @@ class Entity
 public:
 	Entity();
 	bool GetCollides(Entity &other);
-	uint8_t* GetTextureData();
+	uint16_t* GetTextureData();
 	uint16_t GetTextureDataSize();
 
 	void SetPosition(uint8_t x, uint8_t y);
@@ -34,6 +34,6 @@ protected:
 	// There's no data integrity checking between texture data and
 	// entity size. There's also no texture stretching.
 	// Will probably crash if you mismatch texture data size with entity size.
-	uint8_t* m_textureData;
-	uint16_t m_textureDataSize;
+	uint16_t* m_textureData;
+	uint16_t  m_textureDataSize;
 };
