+++ Starship_GalaxyWars/include/ST7735SClient.h
@@ -21,7 +21,7 @@ public:
 	FORCE_INLINE void Initialize();
 	FORCE_INLINE void SetRegion(ScreenRegion& region);
 	FORCE_INLINE void FillCurrentScreenRegion(uint8_t r, uint8_t g, uint8_t b);
-	FORCE_INLINE void FillCurrentScreenRegion(uint8_t* data, uint16_t dataSize);
+	FORCE_INLINE void FillCurrentScreenRegion(uint16_t* data, uint16_t dataSize);
 	FORCE_INLINE void RenderEntity(Entity* entity);
 	FORCE_INLINE bool GetIsInitialized() {return m_initialized;}
